Created a Data Access Layer folder
1. Renamed the RP.Data project into RP.RepositoryPattern
2. Inserted this project into the Data Access Layer folder
3. Necessary namespaces have been renamed
