+++ Data Access Layer/RP.Repository_Pattern/Core/IEmployeeRepository.cs
@@ -1,0 +1,10 @@
+﻿using RP.DataAccess.RepositoryPattern.Entities;
+using System.Collections.Generic;
+
+namespace RP.DataAccess.RepositoryPattern.Core
+{
+    public interface IEmployeeRepository : IRepository<Employee>
+    {
+        IEnumerable<Employee> GetTop5Employees();
+    }
+}
