+++ Data Access Layer/RP.Repository_Pattern/Core/IUnitOfWork.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace RP.DataAccess.RepositoryPattern.Core
+{
+    public interface IUnitOfWork : IDisposable
+    {
+        IEmployeeRepository Employees { get; }
+
+        void Complete();
+    }
+}
