+++ Data Access Layer/RP.Repository_Pattern/EF/RPContext.cs
@@ -1,0 +1,41 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using RP.DataAccess.RepositoryPattern.EF.EntityConfigurations;
+using RP.DataAccess.RepositoryPattern.Entities;
+using System.IO;
+
+namespace RP.DataAccess.RepositoryPattern.EF
+{
+    public class RPContext : DbContext
+    {
+        public RPContext() : base() { }
+
+        public DbSet<Employee> Employees { get; set; }
+
+        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        {
+            var builder = new ConfigurationBuilder()
+                .SetBasePath(Directory.GetCurrentDirectory())
+                .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true);
+
+            IConfigurationRoot configuration = builder.Build();
+
+            optionsBuilder.UseSqlServer(configuration.GetConnectionString("RPContext"));
+        }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.ApplyConfiguration(new EmployeeConfiguration());
+        }
+
+        public new void SaveChanges()
+        {
+            base.SaveChanges();
+        }
+
+        public new void Dispose()
+        {
+            base.Dispose();
+        }
+    }
+}
