+++ Data Access Layer/RP.Repository_Pattern/Entities/Employee.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace RP.DataAccess.RepositoryPattern.Entities
+{
+    public class Employee
+    {
+        public int Id { get; set; }
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public DateTime BirthDate { get; set; }
+    }
+}
