--- RP.Data/Core/IUnitOfWork.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace RP.Data.Core
-{
-    public interface IUnitOfWork : IDisposable
-    {
-        IEmployeeRepository Employees { get; }
-
-        void Complete();
-    }
-}
