--- RP.Data/EF/Repositories/Repository.cs
@@ -1,56 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using RP.Data.Core;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-
-namespace RP.Data.EF.Repositories
-{
-    public class Repository<TEntity> : IRepository<TEntity> where TEntity : class
-    {
-        protected readonly DbContext Context;
-        private readonly DbSet<TEntity> _entities;
-
-        public Repository(DbContext context)
-        {
-            Context = context;
-            _entities = Context.Set<TEntity>();
-        }
-
-        public void Add(TEntity entity)
-        {
-            _entities.Add(entity);
-        }
-
-        public IEnumerable<TEntity> Find(Expression<Func<TEntity, bool>> predicate)
-        {
-            return _entities.Where(predicate);
-        }
-
-        public void AddRange(IEnumerable<TEntity> entities)
-        {
-            _entities.AddRange(entities);
-        }
-
-        public TEntity Get(int id)
-        {
-            return _entities.Find(id);
-        }
-
-        public IEnumerable<TEntity> GetAll()
-        {
-            return _entities.ToList();
-        }
-
-        public void Remove(TEntity entity)
-        {
-            _entities.Remove(entity);
-        }
-
-        public void RemoveRange(IEnumerable<TEntity> entities)
-        {
-            _entities.RemoveRange(entities);
-        }
-    }
-}
