--- RP.Data/Entities/Employee.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace RP.Data.Entities
-{
-    public class Employee
-    {
-        public int Id { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public DateTime BirthDate { get; set; }
-    }
-}
