+++ Tests/RP.IntegrationTests/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "ConnectionStrings": {
+    "RPContext": "Server=(localdb)\\mssqllocaldb;Database=RPDatabase;Trusted_Connection=True;"
+  }
+}
